REFACTOR to use getQueryParams as much as is sane
