+++ lib/client-api/src/story_store.js
@@ -1,7 +1,7 @@
 /* eslint no-underscore-dangle: 0 */
 import { history, document } from 'global';
-import EventEmitter from 'eventemitter3';
 import qs from 'qs';
+import EventEmitter from 'eventemitter3';
 import memoize from 'memoizerific';
 import debounce from 'lodash.debounce';
 import { stripIndents } from 'common-tags';
@@ -11,6 +11,7 @@ import { logger } from '@storybook/client-logger';
 import { toId } from '@storybook/router/utils';
 
 import pathToId from './pathToId';
+import { getQueryParams } from './queryparams';
 
 // TODO: these are copies from components/nav/lib
 // refactor to DRY
@@ -65,7 +66,7 @@ export default class StoryStore extends EventEmitter {
     this.on(Events.STORY_INIT, () => {
       let storyId = this.getIdOnPath();
       if (!storyId) {
-        const query = qs.parse(document.location.search, { ignoreQueryPrefix: true });
+        const query = getQueryParams();
         storyId = getIdFromLegacyQuery(query);
         if (storyId) {
           const { path, selectedKind, selectedStory, ...rest } = query;
@@ -83,7 +84,7 @@ export default class StoryStore extends EventEmitter {
   // NEW apis
 
   getIdOnPath = () => {
-    const { id } = qs.parse(document.location.search, { ignoreQueryPrefix: true });
+    const { id } = getQueryParams();
     return id;
   };
